Changes related to Ribbon implementation
