+++ currency-conversion-service/src/main/java/com/microservice/currencyconversionservice/CurrencyConversionController.java
@@ -1,0 +1,29 @@
+package com.microservice.currencyconversionservice;
+
+import java.math.BigDecimal;
+import java.util.HashMap;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestClientException;
+import org.springframework.web.client.RestTemplate;
+
+@RestController
+public class CurrencyConversionController {
+	@Autowired
+	private CurrencyExchangeServiceProxy proxy;
+
+	@GetMapping("/currency-converter-feign/from/{from}/to/{to}/quantity/{quantity}")
+	public CurrencyConversionBean getConvertCurrencyFeign(@PathVariable String from, @PathVariable String to,
+			@PathVariable BigDecimal quantity) {
+		
+			CurrencyConversionBean body = proxy.retrieveExchangeValue(from, to);
+			return new CurrencyConversionBean(body.getId(), from, to, body.getConversionMultiple(), quantity,
+					quantity.multiply(body.getConversionMultiple()), body.getPort());
+		
+
+	}
+}
