+++ currency-conversion-service/src/main/java/com/microservice/currencyconversionservice/CurrencyConversionServiceApplication.java
@@ -1,0 +1,15 @@
+package com.microservice.currencyconversionservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients("com.microservice.currencyconversionservice")
+public class CurrencyConversionServiceApplication 
+{
+public static void main(String[] args) 
+{
+SpringApplication.run(CurrencyConversionServiceApplication.class, args);
+}
+}
