--- currency-conversion-service/src/main/java/javatpoint/microservice/currencyconversionservice/CurrencyConversionServiceApplication.java
@@ -1,15 +1,0 @@
-package javatpoint.microservice.currencyconversionservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients("javatpoint.microservice.currencyconversionservice")
-public class CurrencyConversionServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CurrencyConversionServiceApplication.class, args);
-	}
-
-}
