--- currency-conversion-service/src/main/java/javatpoint/microservice/currencyconversionservice/CurrencyExchangeServiceProxy.java
@@ -1,13 +1,0 @@
-package javatpoint.microservice.currencyconversionservice;
-
-import java.math.BigDecimal;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
-@FeignClient(name="currency-exchange", url="localhost:8000")
-public interface CurrencyExchangeServiceProxy {
-	@GetMapping("/currency-exchange/from/{from}/to/{to}")
-	public CurrencyConversionBean convertCurrency(@PathVariable String from, @PathVariable String to);
-}
