+++ currency-conversion-service/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 spring.application.name=currency-conversion-service  
 server.port=8100 
+currency-exchange-service.ribbon.listOfServers=http://localhost:8000,http://localhost:8001,http://localhost:8002
